Refactored getAll.
